# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"],"security":[{"bearer":[]}]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":false,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/auth":{"get":{"operationId":"AuthController_checkUserController","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]},"post":{"operationId":"AuthController_loginController","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/produk":{"post":{"operationId":"ProdukController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProdukDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Produk"],"security":[{"bearer":[]}]},"get":{"operationId":"ProdukController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"barcode","required":false,"in":"query","schema":{"type":"string"}},{"name":"nama_produk","required":false,"in":"query","schema":{"type":"string"}},{"name":"deskripsi_produk","required":false,"in":"query","schema":{"type":"string"}},{"name":"harga_beli","required":false,"in":"query","schema":{"type":"number"}},{"name":"harga_jual","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponProdukDto"}}}}},"tags":["Produk"],"security":[{"bearer":[]}]}},"/produk/{id}":{"get":{"operationId":"ProdukController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Produk"}}}}},"tags":["Produk"],"security":[{"bearer":[]}]},"patch":{"operationId":"ProdukController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProdukDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Produk"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProdukController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Produk"}}}}},"tags":["Produk"],"security":[{"bearer":[]}]}},"/produk/produk-manual-query":{"post":{"operationId":"ProdukController_produkManualQuery","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProdukManualQueryDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Produk"],"security":[{"bearer":[]}]}},"/konsumen":{"post":{"operationId":"KonsumenController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateKonsumanDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Konsumen"],"security":[{"bearer":[]}]},"get":{"operationId":"KonsumenController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"nama_konsumen","required":false,"in":"query","schema":{"type":"string"}},{"name":"alamat_konsumen","required":false,"in":"query","schema":{"type":"string"}},{"name":"email_konsumen","required":false,"in":"query","schema":{"type":"string"}},{"name":"no_hp_konsumen","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseKonsumenDto"}}}}},"tags":["Konsumen"],"security":[{"bearer":[]}]}},"/konsumen/{id}":{"get":{"operationId":"KonsumenController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Konsuman"}}}}},"tags":["Konsumen"],"security":[{"bearer":[]}]},"patch":{"operationId":"KonsumenController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateKonsumanDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Konsumen"],"security":[{"bearer":[]}]},"delete":{"operationId":"KonsumenController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Konsuman"}}}}},"tags":["Konsumen"],"security":[{"bearer":[]}]}},"/rekening":{"post":{"operationId":"RekeningController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRekeningDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Rekening"],"security":[{"bearer":[]}]},"get":{"operationId":"RekeningController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"nama_rekening","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponRekeningDto"}}}}},"tags":["Rekening"],"security":[{"bearer":[]}]}},"/rekening/{id}":{"get":{"operationId":"RekeningController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rekening"}}}}},"tags":["Rekening"],"security":[{"bearer":[]}]},"patch":{"operationId":"RekeningController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRekeningDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Rekening"],"security":[{"bearer":[]}]},"delete":{"operationId":"RekeningController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Rekening"}}}}},"tags":["Rekening"],"security":[{"bearer":[]}]}},"/penjualan":{"post":{"operationId":"PenjualanController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePenjualanDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Penjualan"],"security":[{"bearer":[]}]},"get":{"operationId":"PenjualanController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"no_faktur","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponProdukDto"}}}}},"tags":["Penjualan"],"security":[{"bearer":[]}]}},"/penjualan/{id}":{"get":{"operationId":"PenjualanController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Penjualan"],"security":[{"bearer":[]}]},"patch":{"operationId":"PenjualanController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePenjualanDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Penjualan"],"security":[{"bearer":[]}]},"delete":{"operationId":"PenjualanController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Penjualan"}}}}},"tags":["Penjualan"],"security":[{"bearer":[]}]}}},"info":{"title":"OPEN-API POS","description":"Documentasi untuk api point of sale","version":"1.3","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"nama_user":{"type":"string","minLength":8,"maxLength":64},"email":{"type":"string","minLength":6,"maxLength":32},"username":{"type":"string","minLength":8,"maxLength":32},"password":{"type":"string","minLength":8,"maxLength":32}},"required":["nama_user","email","username","password"]},"Konsuman":{"type":"object","properties":{"id":{"type":"number"},"nama_konsumen":{"type":"string"},"alamat_konsumen":{"type":"string"},"email_konsumen":{"type":"string"},"no_hp_konsumen":{"type":"string"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","nama_konsumen","alamat_konsumen","email_konsumen","no_hp_konsumen","create_at","update_at","user"]},"Rekening":{"type":"object","properties":{"id":{"type":"number"},"nama_rekening":{"type":"string"},"keterangan_rekening":{"type":"string"},"type_rekening":{"type":"string"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","nama_rekening","keterangan_rekening","type_rekening","create_at","update_at","user"]},"User":{"type":"object","properties":{"id":{"type":"number"},"nama_user":{"type":"string"},"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"produk":{"$ref":"#/components/schemas/Produk"},"konsumen":{"$ref":"#/components/schemas/Konsuman"},"rekening":{"$ref":"#/components/schemas/Rekening"}},"required":["id","nama_user","email","username","password","create_at","update_at","produk","konsumen","rekening"]},"Produk":{"type":"object","properties":{"id":{"type":"number"},"barcode":{"type":"string"},"nama_produk":{"type":"string"},"deskripsi_produk":{"type":"string"},"harga_beli":{"type":"number"},"harga_jual":{"type":"number"},"foto":{"type":"string"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","barcode","nama_produk","deskripsi_produk","harga_beli","harga_jual","foto","create_at","update_at","user"]},"UpdateUserDto":{"type":"object","properties":{"id":{"type":"number"},"nama_user":{"type":"string","minLength":8,"maxLength":64},"email":{"type":"string","minLength":6,"maxLength":32},"username":{"type":"string","minLength":8,"maxLength":32},"password":{"type":"string","minLength":8,"maxLength":32}}},"AuthDto":{"type":"object","properties":{"username":{"type":"string","minLength":8,"maxLength":32},"password":{"type":"string","minLength":8,"maxLength":32}},"required":["username","password"]},"CreateProdukDto":{"type":"object","properties":{"barcode":{"type":"string"},"nama_produk":{"type":"string"},"deskripsi_produk":{"type":"string"},"harga_beli":{"type":"number"},"harga_jual":{"type":"number"},"foto":{"type":"string","format":"binary"}},"required":["barcode","nama_produk","deskripsi_produk","harga_beli","harga_jual","foto"]},"ProdukDto":{"type":"object","properties":{"id":{"type":"number"},"barcode":{"type":"string"},"nama_produk":{"type":"string"},"deskripsi_produk":{"type":"string"},"harga_beli":{"type":"number"},"harga_jual":{"type":"number"},"foto":{"type":"string","format":"binary"}},"required":["id","barcode","nama_produk","deskripsi_produk","harga_beli","harga_jual","foto"]},"ResponProdukDto":{"type":"object","properties":{"total":{"type":"number"},"pages":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ProdukDto"}}},"required":["total","pages","data"]},"UpdateProdukDto":{"type":"object","properties":{"id":{"type":"number"},"barcode":{"type":"string"},"nama_produk":{"type":"string"},"deskripsi_produk":{"type":"string"},"harga_beli":{"type":"number"},"harga_jual":{"type":"number"},"foto":{"type":"string","format":"binary"}}},"ProdukManualQueryDto":{"type":"object","properties":{"condition":{"type":"object","properties":{"barcode":{"required":true,"type":"string"}}},"skip":{"type":"number","default":0},"limit":{"type":"number","default":10},"enable_count":{"type":"number","default":1},"enable_manual_relation_user":{"type":"number","default":1}},"required":["condition"]},"CreateKonsumanDto":{"type":"object","properties":{"nama_konsumen":{"type":"string"},"alamat_konsumen":{"type":"string"},"email_konsumen":{"type":"string"},"no_hp_konsumen":{"type":"string"}},"required":["nama_konsumen","alamat_konsumen","email_konsumen","no_hp_konsumen"]},"KonsumenDto":{"type":"object","properties":{"id":{"type":"number"},"nama_konsumen":{"type":"string"},"alamat_konsumen":{"type":"string"},"email_konsumen":{"type":"string"},"no_hp_konsumen":{"type":"string"}},"required":["id","nama_konsumen","alamat_konsumen","email_konsumen","no_hp_konsumen"]},"ResponseKonsumenDto":{"type":"object","properties":{"total":{"type":"number"},"pages":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/KonsumenDto"}}},"required":["total","pages","data"]},"UpdateKonsumanDto":{"type":"object","properties":{"id":{"type":"number"},"nama_konsumen":{"type":"string"},"alamat_konsumen":{"type":"string"},"email_konsumen":{"type":"string"},"no_hp_konsumen":{"type":"string"}}},"CreateRekeningDto":{"type":"object","properties":{"nama_rekening":{"type":"string"},"keterangan_rekening":{"type":"string"},"type_rekening":{"type":"string"}},"required":["nama_rekening","keterangan_rekening","type_rekening"]},"RekeningDto":{"type":"object","properties":{"id":{"type":"number"},"nama_rekening":{"type":"string"},"keterangan_rekening":{"type":"string"},"type_rekening":{"type":"string"}},"required":["id","nama_rekening","keterangan_rekening","type_rekening"]},"ResponRekeningDto":{"type":"object","properties":{"total":{"type":"number"},"pages":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/RekeningDto"}}},"required":["total","pages","data"]},"UpdateRekeningDto":{"type":"object","properties":{"id":{"type":"number"},"nama_rekening":{"type":"string"},"keterangan_rekening":{"type":"string"},"type_rekening":{"type":"string"}}},"KonsumenId":{"type":"object","properties":{"id":{"type":"number"}},"required":["id"]},"ProdukIdDto":{"type":"object","properties":{"id":{"type":"number"}},"required":["id"]},"PenjualanItemDto":{"type":"object","properties":{"jumlah_jual":{"type":"number","minimum":1},"harga_jual":{"type":"number"},"potongan":{"type":"number"},"produk":{"$ref":"#/components/schemas/ProdukIdDto"}},"required":["jumlah_jual","harga_jual","potongan","produk"]},"RekeningIdDto":{"type":"object","properties":{"id":{"type":"number"}},"required":["id"]},"PenjualanBayarDto":{"type":"object","properties":{"tanggal_bayar":{"format":"date-time","type":"string"},"jumlah_bayar":{"type":"number"},"rekening":{"$ref":"#/components/schemas/RekeningIdDto"}},"required":["tanggal_bayar","jumlah_bayar","rekening"]},"CreatePenjualanDto":{"type":"object","properties":{"no_faktur":{"type":"string"},"tanggal":{"format":"date-time","type":"string"},"konsumen":{"$ref":"#/components/schemas/KonsumenId"},"item":{"type":"array","items":{"$ref":"#/components/schemas/PenjualanItemDto"}},"bayar":{"type":"array","items":{"$ref":"#/components/schemas/PenjualanBayarDto"}}},"required":["no_faktur","tanggal","konsumen","item","bayar"]},"UpdatePenjualanDto":{"type":"object","properties":{"id":{"type":"number"},"no_faktur":{"type":"string"},"tanggal":{"format":"date-time","type":"string"},"konsumen":{"$ref":"#/components/schemas/KonsumenId"},"item":{"type":"array","items":{"$ref":"#/components/schemas/PenjualanItemDto"}},"bayar":{"type":"array","items":{"$ref":"#/components/schemas/PenjualanBayarDto"}}}},"PenjualanBayar":{"type":"object","properties":{"id":{"type":"number"},"tanggal_bayar":{"format":"date-time","type":"string"},"jumlah_bayar":{"type":"number"},"penjualan":{"$ref":"#/components/schemas/Penjualan"},"rekening":{"$ref":"#/components/schemas/Rekening"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","tanggal_bayar","jumlah_bayar","penjualan","rekening","create_at","update_at","user"]},"Penjualan":{"type":"object","properties":{"id":{"type":"number"},"no_faktur":{"type":"string"},"tanggal":{"format":"date-time","type":"string"},"total_transaksi":{"type":"number"},"total_potongan":{"type":"number"},"total_bayar":{"type":"number"},"konsumen":{"$ref":"#/components/schemas/Konsuman"},"item":{"type":"array","items":{"$ref":"#/components/schemas/PenjualanItem"}},"bayar":{"type":"array","items":{"$ref":"#/components/schemas/PenjualanBayar"}},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","no_faktur","tanggal","total_transaksi","total_potongan","total_bayar","konsumen","item","bayar","create_at","update_at","user"]},"PenjualanItem":{"type":"object","properties":{"id":{"type":"number"},"jumlah_jual":{"type":"number"},"harga_jual":{"type":"number"},"potongan":{"type":"number"},"penjualan":{"$ref":"#/components/schemas/Penjualan"},"produk":{"$ref":"#/components/schemas/Produk"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","jumlah_jual","harga_jual","potongan","penjualan","produk","create_at","update_at","user"]}}}}